made camera cutout wider
